fix(trusty-mpm): repoint dangling test pointers, regenerate tm-capabilities
Two CI checks I did not run locally were red. Both are fallout from the
#4286 retirement.

1. Doc-comment pointer lint: 16 pointers cited test names removed or
   renamed by the retirement. Each was repointed at the test that now
   covers the behaviour, not deleted to silence the lint.

   Two of them named behaviours that STILL EXIST but had lost their only
   coverage when the retired-file tests went:

   - `MEMORY_OVERRIDE_HEADING` slotting — still live in `assemble_sections`,
     reachable via a CLAUDE.md MEMORY block on the roster-absent path.
   - `warn_unapplied` for sections the string assembly cannot address —
     still called; its test went with the #4399 cluster.

   Deleting those two pointers would have hidden untested live behaviour,
   so coverage was restored first:
   `a_named_memory_override_is_slotted_on_the_roster_absent_path` and
   `unaddressable_sections_are_reported_unapplied_on_the_roster_absent_path`.

2. tm-capabilities drift: the new `legacy_overrides` doctor check was not
   reflected. Regenerated with `tm generate capabilities`; verified the
   output actually carries the check (position 15, count 28 -> 29) rather
   than assuming.

Adding the two tests took `instruction_overrides.rs` to 538 SLOC, over the
500 production cap, so its inline test module is extracted to
`instruction_overrides_tests.rs` via `#[path]` — the same split
`claude_md_sections.rs` already uses.

Refs #4286
